@@ -4,13 +4,15 @@
   "description": "Node.js, Express.js Mongoose, MongoDB Backend",
   "main": "src/index.ts",
   "scripts": {
-    "dev": "ts-node-dev --respawn src/server.ts",
+    "dev": "ts-node-dev --respawn src/server.ts ",
     "heroku-prebuild": "npm install --only=dev",
-    "start": "npm install && npx tsc --project ./ && node build/server.js",
+    "start": "npm install && npx tsc --project ./ && node build/src/server.js",
     "test": "jest --coverage",
-    "build": "tsc"
+    "build": "tsc",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix"
   },
-  "author": "Ken Beaudin, Dmitry Kulakov, Abdul Shabbir & Enrick Anderson Ong",
+  "author": "Ken Beaudin, Dmitry Kulakov, Abdul Shabbir & Enrick Andersen Ong",
   "license": "MIT",
   "type": "commonjs",
   "dependencies": {
@@ -19,10 +21,10 @@
     "@types/express": "^4.17.2",
     "@types/jest": "^29.0.3",
     "@types/jsonwebtoken": "^8.5.9",
-    "@types/node": "^18.7.5",
     "@types/supertest": "^2.0.12",
     "@types/validator": "^13.7.10",
     "bcrypt": "^5.1.0",
+    "@types/morgan": "^1.9.3",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
@@ -37,7 +39,23 @@
     "tls": "^0.0.1",
     "ts-jest": "^29.0.1",
     "ts-node-dev": "^2.0.0",
-    "typescript": "^4.8.4",
-    "validator": "^13.7.0"
+    "validator": "^13.7.0",
+    "@babel/core": "^7.20.5",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/preset-react": "^7.18.6",
+    "@types/node": "^18.11.10",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
+    "eslint": "^8.28.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-wesbos": "^3.1.4",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "prettier": "^2.8.0",
+    "typescript": "^4.9.3"
+  },
+  "devDependencies": {
   }
 }
